cart-added

+++ src/main/java/io/grocery/backend/entity/Admin.java
@@ -7,21 +7,27 @@ import jakarta.persistence.Id;
 
 @Entity
 public class Admin {
-	  @Id
-	  @GeneratedValue(strategy=GenerationType.AUTO)
-	private Integer id;
-private String email;
-public String getEmail() {
-	return email;
-}
-public void setEmail(String email) {
-	this.email = email;
-}
-public String getPassword() {
-	return password;
-}
-public void setPassword(String password) {
-	this.password = password;
-}
-private String password;
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private Long id;
+	
+	private String email;
+
+	public String getEmail() {
+		return email;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
+
+	public String getPassword() {
+		return password;
+	}
+
+	public void setPassword(String password) {
+		this.password = password;
+	}
+
+	private String password;
 }
