+++ src/main/java/io/grocery/backend/entity/Products.java
@@ -15,12 +15,12 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Entity
 public class Products {
-	  @Id
-	  @GeneratedValue(strategy=GenerationType.AUTO)
-	private Integer id;
-private String title;
-private Integer price;
-private String description;
-private String image;
-private Integer quantity;
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private Long id;
+	private String title;
+	private Integer price;
+	private String description;
+	private String image;
+	private Integer quantity;
 }
